console: add cd and per-process cwd; bind scheduler current-process accessors in mc
Console now tracks a working directory per process: the prompt shows the cwd
name (/ at root), ls/cat resolve relative to proc->cwd, and a new cd built-in
changes it. cd is dispatched directly rather than forked so the change lands in
the console process itself.

Scheduler gains an mc binding layer (kernel/scheduler.mc) that @extern-binds the
C core and implements scheduler_get/set_current_process in mc; scheduler.c routes
all current-process access through these accessors and exposes current/idle_ctx.

libmc: add a growable string (array<uint8> specialization), refactor set_entry to
extend pair and dict_entry to extend set_entry, add dict iteration, and rename the
for-in protocol from iter/next to type-prefixed array_it/set_it/dict_it + *_next.

Widen fat32_mount to int64 across the C and mc declarations.

+++ src/fs/fat32.c
@@ -7,6 +7,7 @@
 #include <fs/vfs.h>
 #include <mm/heap.h>
 #include <stdbool.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <string.h>
 #include <uchar.h>
@@ -529,7 +530,7 @@ static int _read_fat_table(char *pathname, struct fat32_bs_info *bs_info, uint8_
     return bs_info->table_size_32;
 }
 
-int fat32_mount(char *device_path, char *mountpoint) {
+int64_t fat32_mount(char *device_path, char *mountpoint) {
     int status;
     uint8_t bs[512];
     char _mountpoint[50] = {0};
